# Optimize reverb by removing stray `volatile` from the `undenormalize` function signature.