--- a/core/math/audio_frame.h
+++ b/core/math/audio_frame.h
@@ -33,7 +33,7 @@
 #include "core/math/vector2.h"
 #include "core/typedefs.h"
 
-static inline float undenormalize(volatile float f) {
+static _FORCE_INLINE_ float undenormalize(float f) {
 	union {
 		uint32_t i;
 		float f;